requirements.txt
python-telegram-bot>=22.2
anthropic>=0.7.0
python-dotenv>=1.0.0
